tfc_project/requirements.txt
pip==22.3
django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
python-dateutil==2.8.2
Pillow==9.3.0
pytz==2022.6
asgiref==3.5.2
sqlparse==0.4.3
setuptools==63.2.0
packaging==21.3
pyparsing==3.0.9
PyJWT==2.6.0
geopy==2.3.0
haversine==2.7.0
datetime==4.7
pathlib==1.0.1
django-cors-headers==3.13.0